All classes and file 'most_common_words.txt' are moved to the appropriate folders
