--- src/ua/com/javarush/vika_kovalenko/cryptoanalyser/Main.java
@@ -1,7 +1,0 @@
-package ua.com.javarush.vika_kovalenko.cryptoanalyser;
-
-public class Main {
-    public static void main(String[] args) {
-        Dialog.dialogWithUser();
-    }
-}
