+++ src/ua/com/javarush/vika_kovalenko/cryptoanalyser/files/FileChecking.java
@@ -1,4 +1,7 @@
-package ua.com.javarush.vika_kovalenko.cryptoanalyser;
+package ua.com.javarush.vika_kovalenko.cryptoanalyser.files;
+
+import ua.com.javarush.vika_kovalenko.cryptoanalyser.model.Artifacts;
+import ua.com.javarush.vika_kovalenko.cryptoanalyser.exception.FileProcessingException;
 
 import java.io.File;
 import java.io.IOException;
