+++ src/ua/com/javarush/vika_kovalenko/cryptoanalyser/model/Alphabet.java
@@ -1,4 +1,4 @@
-package ua.com.javarush.vika_kovalenko.cryptoanalyser;
+package ua.com.javarush.vika_kovalenko.cryptoanalyser.model;
 
 public class Alphabet {
